src/database/db_mysql.py
from decouple import config

import pymysql
import traceback


def get_connecction():
    try:
        return pymysql.connect(
            host=config('MYSQL_HOST'),
            user=config('MYSQL_USER'),
            password=config('MYSQL_PASSWORD'),
            db=config('MYSQL_DB')
        )
    except Exception as ex:
        return f'Error de conexion: ' , ex